% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cts.R
\name{add_cts}
\alias{add_cts}
\title{Aggregator function
Adds each local cross table to form a global cross tabulation.}
\usage{
add_cts(nodes, master)
}
\arguments{
\item{nodes}{list of node specific cross table}

\item{master}{list containing output from `init_formula` and
`variable_categories`.}
}
\value{
Global cross tabulation.
}
\description{
Aggregator function
Adds each local cross table to form a global cross tabulation.
}
