% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.GLM.R
\name{as.GLM}
\alias{as.GLM}
\title{Convert the federated GLM output to a GLM object}
\usage{
as.GLM(obj, data = NULL)
}
\arguments{
\item{obj}{output of the federated GLM}

\item{data}{...}
}
\value{
GLM object
}
\description{
Convert the federated GLM output to a GLM object
}
