% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_types.R
\name{assign_types}
\alias{assign_types}
\title{Assign numeric or factor type to columns specified.}
\usage{
assign_types(data, types)
}
\arguments{
\item{data}{dataframe}

\item{types}{containing the types to set to the columns}
}
\value{
formatted dataframe
}
\description{
Assign numeric or factor type to columns specified.
}
