% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_na.R
\name{comb_na}
\alias{comb_na}
\title{Combine NA lengths.}
\usage{
comb_na(node_nas, col)
}
\arguments{
\item{node_nas}{a list of node specific missing values per column}

\item{col}{Should be supplied by researcher as a vector of strings
representing the column names they think is present in the data}
}
\value{
a vector of missing values per column
}
\description{
Each datastation (node) sends back the number of missing value per column,
this function simply aggregates it.
}
