% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_range.R
\name{comb_range}
\alias{comb_range}
\title{Function to combine range(s) of the columns.}
\usage{
comb_range(node.range, col)
}
\arguments{
\item{node.range}{these are the node specific range(s) per column. This is
a list of either a vector with 2 distinct values (min, max) or a table
for factor data of the number of unique instances of each factor level.}

\item{col}{Should be supplied by researcher as a vector of strings
representing the column names they think is present in the data}
}
\value{
list of combined ranges.
}
\description{
This function combines the ranges sent back by each datastation (node)
based on the type of range. Here it can be two unique cases: a table for
factor data or a numeric/integer in which case the simple min/max of each
datasation per unique column is aggregated.
}
