% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_sums.R
\name{comb_sums}
\alias{comb_sums}
\title{Combine Sums}
\usage{
comb_sums(node.list.to.sum, col)
}
\arguments{
\item{node.list.to.sum}{is the node specific sum per column. This is a list.}

\item{col}{Should be supplied by researcher as a vector of strings
representing the column names they think is present in the data}
}
\value{
named vector of aggregated sums.
}
\description{
Function to aggregate the datastation (node) specific sums per column.
}
