% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.combined.ties.R
\name{compute.combined.ties}
\alias{compute.combined.ties}
\title{Compute/count the *global* number of ties at each time point using the
tie-counts from the individual sites.}
\usage{
compute.combined.ties(Ds)
}
\description{
Params:
  Ds: list of *local* tie-counts, indexed by site nr
}
\details{
Return:
  numeric vector with tie-counts and named index with unique
  event times
}
