% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.derivatives.R
\name{compute.derivatives}
\alias{compute.derivatives}
\title{Compute the primary derivative (vector of length beta)}
\usage{
compute.derivatives(z_hat, D_all, aggregates)
}
\description{
Compute the primary derivative (vector of length beta)
}
