% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxfit.R
\name{coxfit}
\alias{coxfit}
\title{Server side, collects the coxfit run by vtg.coxph}
\usage{
coxfit(fit, transform = "identity")
}
\value{
return list with beta coefficients and betavar
}
\description{
This only looks at the relevant output from vtg.coxph, this means that it
looks at the coefficients (Beta) and the variance-covariance matrix
(Betavar)
}
