% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dchisq.R
\name{dchisq}
\alias{dchisq}
\title{Federated Chisq Test.}
\usage{
dchisq(
  client,
  col,
  threshold = 5L,
  probs = NULL,
  organizations_to_include = NULL
)
}
\arguments{
\item{client}{vtg::Client instance provided by node (datastation).}

\item{col}{Can by single column name or N column name. If `2` column names,
executes Chisq on Contingency table. Warning: Sends frequency distribution.}

\item{threshold}{Disclosure check. Default is 5, if number of counts in
any cell is less than `threshold` the function stops and returns an error
message.}

\item{probs}{These are the probabilities needed. Default is `NULL` however
the data-owner/researcher can supply their own. The length of which has to
correspond to the "total" length of all combined dataset for given
column(s).}
}
\description{
This version has built in `threshold` parameter that checks if any counts
are less than tolerance. Default is 5. Can go lower (to 1). Up to data-owner.
}
\author{
Hasan Alradhi

Matteo Cellamare

Frank Martin
}
