% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcoxph.mock.R
\name{dcoxph.mock}
\alias{dcoxph.mock}
\title{Run the distributed CoxPH algorithm locally}
\usage{
dcoxph.mock(df, input_data, splits = 5)
}
\description{
Splits the provided data frame in `splits` equal parts and runs
`dcoxph()`.
}
\details{
Params:
  df: data frame containing the *full* dataset
  expl_vars: list of explanatory variables (covariates) to use
  time_col: name of the column that contains the event/censor times
  censor_col: name of the colunm that explains whether an event occured or
              the patient was censored
  splits: number of parts to split the data set in

Return:
  data.frame with beta, p-value and confidence interval for each explanatory
  variable.
}
