% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcoxph.R
\name{dcoxph}
\alias{dcoxph}
\title{Run the distributed CoxPH algorithm.}
\usage{
dcoxph(
  client,
  expl_vars,
  time_col,
  censor_col,
  types = NULL,
  organizations_to_include = NULL,
  subset_rules = NULL
)
}
\description{
Params:
  client: ptmclient::Client instance.
  input_data: input_data containing
     expl_vars: list of explanatory variables (covariates) to use
     time_col: name of the column that contains the event/censor times
     censor_col: name of the column that explains whether an event occurred
                 or the patient was censored
     organizations_to_include: either NULL meaning all  participating
                               organisations or select organisation ids;
                               must be list of id(s).
}
\details{
Return:
  RDS with beta, p-value and confidence interval for each explanatory
  variable.
}
