% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct.R
\name{dct}
\alias{dct}
\title{Federated Cross Tabulation}
\usage{
dct(client, f)
}
\arguments{
\item{client}{vtg::Client instance, provided by the node}

\item{f}{an object of class formula}
}
\value{
Federated Cross Table object.
}
\description{
Federated Cross Tabulation
}
\author{
Alradhi, H.

Cellamare, M.
}
