% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dglm.R
\name{dglm}
\alias{dglm}
\title{Run the distributed GLM algorithm.}
\usage{
dglm(
  client,
  formula,
  dstar = NULL,
  types = NULL,
  family = gaussian,
  tol = 1e-08,
  maxit = 25
)
}
\description{
Params:
   client: ptmclient::Client instance.
   formula: dependant_variable ~ explanatory_variable(i) + ...
   types: types of the columns that are used in the formula
   family: to this up it uses the Gaussian as this is the default value
   tol: tolerance level
   maxit: maximum number of iterations the function is allowed to cycle up to
}
