% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsummary.R
\name{dsummary}
\alias{dsummary}
\title{Federated Summary algorithm.}
\usage{
dsummary(client, col, threshold = 5L, organizations_to_include = NULL)
}
\arguments{
\item{client}{vtg::Client instance provided by node (datastation).}

\item{col}{Can by single column name or N column name.}

\item{threshold}{Disclosure check. Default is 5, if number of counts in
any cell is less than `threshold` the function stops and returns an error
message.}
}
\description{
This version has built in `threshold` parameter that checks if any counts
are less than tolerance. Default is 5. Can go lower (to 1). Up to data-owner.
}
\author{
Hasan Alradhi

Matteo Cellamare

Frank Martin
}
