% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsurvdiff.R
\name{dsurvdiff}
\alias{dsurvdiff}
\title{Federated survdiff}
\usage{
dsurvdiff(client, formula, tmax = NA, timepoints = NULL)
}
\arguments{
\item{client}{vtg::Client instance, provided by the node}

\item{formula}{an object of class formula (or one that can be coerced
to that class: a symbolic description of the model to be fitted.
E.g.: dependant_variable ~ explanatory_variable(i) + ...}

\item{timepoints}{time points to calculate KM (bins instead of individual time point)}
}
\value{
for each strata returns a table with (n events median 0.95LCL 0.95UCL) KM plot and ...
}
\description{
Federated survdiff
}
\author{
Cellamare, M.

Alradhi, H.

Martin, F.
}
