% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsurvfit.R
\name{dsurvfit}
\alias{dsurvfit}
\title{Federated survfit (Kaplan Meier)}
\usage{
dsurvfit(
  client,
  formula,
  conf.int = 0.95,
  conf.type = "log",
  tmax = NA,
  timepoints = NULL,
  plotCI = F,
  organizations_to_include = NULL,
  subset_rules = NULL
)
}
\arguments{
\item{client}{vtg::Client instance, provided by the node}

\item{formula}{an object of class formula (or one that can be coerced
to that class: a symbolic description of the model to be fitted.
E.g.: dependant_variable ~ explanatory_variable(i) + ...}

\item{conf.int}{confidence interval coverage (95% default)}

\item{conf.type}{type of confidence interval (log, identity, log-log)}

\item{timepoints}{time points to calculate KM (bins instead of individual
time point)}

\item{plotCI}{True if the researcher wants to plot Confidence Interval for
the KM curves}

\item{organizations_to_include}{either NULL meaning all participating
organizations or select organizations ids; must be list of id(s)}
}
\value{
for each strata returns a table with (n events median 0.95LCL
0.95UCL) KM plot and ...
}
\description{
Federated survfit (Kaplan Meier)
}
\author{
Cellamare, M.

Alradhi, H.

Martin, F.
}
