% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example1}
\alias{example1}
\alias{example2}
\alias{example3}
\title{Example dataset for GLM}
\format{
A data_frame:
 \describe{
    \item{X}{id}
    \item{num_awards}{outcome variable and indicates the number of
      awards earned by students}
    \item{prog}{categorical predictor variable with three levels
      indicating the type of program}
    \item{math}{Continuous predictor variable and represents students’
      scores}
  }

An object of class \code{data.frame} with 64 rows and 5 columns.

An object of class \code{data.frame} with 70 rows and 5 columns.
}
\source{
https://stats.idre.ucla.edu/stat/data/poisson_sim.csv
}
\usage{
example1

example2

example3
}
\description{
Slightly modified dataset from `poisson_sim.csv`, see
 the source bellow
}
\examples{
data(example1)
data(example2)
data(example3)


}
\keyword{datasets}
