% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_data.R
\name{extend_data}
\alias{extend_data}
\title{version: 1.0
Extend the data with new variables}
\usage{
extend_data(data)
}
\arguments{
\item{data}{Data provided by client.}
}
\value{
Original dataframe extended with new columns.
}
\description{
This function enriches the datasets at the nodes with more variable which in
turn can be used in analyses. Created together with Laura Botta (INT).
This function will be called at the start of the analysis, and executed on
the data of every node. Has to be implemented in every RPC call.
}
