% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{assign types to all columns from the data-frame}
\usage{
format_data(data, types)
}
\arguments{
\item{data}{dataframe}

\item{types}{containing the types to set to the columns}
}
\value{
formatted dataframe
}
\description{
assign types to all columns from the data-frame
}
