% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_family.R
\name{get_family}
\alias{get_family}
\title{translates family input to R family object}
\usage{
get_family(family, dstar = NULL, data = NULL)
}
\arguments{
\item{family}{family input}
}
\value{
R family object
}
\description{
translates family input to R family object
}
