% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm.R
\name{glmm}
\alias{glmm}
\title{Federated Generalized Mixed Effect Function.}
\usage{
glmm(client, params, local_eval, family, nAGQ, formula)
}
\arguments{
\item{params}{List of starting values for the Random Effect Term and the
Fixed Effect Term(s). Note that one should invoke the list in the following
manner: params = list(theta = x, fixef = y) where `x` is a single value, and
`y` can be a vector of starting values depending on how many fixed effects
are included in the model.}

\item{local_eval}{String, RPC call to run a single iteration of
the `lme4::glmer` function. This evaluates the deviance at the supplied set
of starting values given by the `params` parameter via the Adaptive Gauss
Hermite Quadrature Scheme.}

\item{family}{Family type, if non is supplied then Gaussian Family is used.}

\item{nAGQ}{Integer Scalar of the number of points per axis for evaluating
the Adaptive Gauss Hermite Quadrature approximation to the log-likelihood.
The default number we take is 10, but this can go up to 25 or as few as 1.
The latter will simply be the Laplace Approximation of the Deviance.}

\item{formula}{A two-sided linear formula object describing both the fixed
effects and random effect parts of the model. The response parameter must be
on the left hand side of a '~' operator and the independent variables that
follow must be separated with a '+' operator. The random effect term must
be distinguished by a vertical bar '|'.}

\item{vtg::Client}{instance, provided by the node}
}
\value{
A list object from which
}
\description{
Collects all the partial values of the deviance from each site via
vtg.glmm::concatenate_results then
}
