% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glob_mean.R
\name{glob_mean}
\alias{glob_mean}
\title{Global Mean.}
\usage{
glob_mean(glob.sums, glob.lens, col)
}
\arguments{
\item{glob.sums}{is the global sum per column.}

\item{glob.lens}{is the global lengths of each column.}

\item{col}{Should be supplied by researcher as a vector of strings
representing the column names they think is present in the data}
}
\value{
named vector of global mean per column.
}
\description{
Function to compute the global mean from aggregated sums and lengths of each
column.
}
