% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glob_var.R
\name{glob_var}
\alias{glob_var}
\title{Global Variance.}
\usage{
glob_var(glob.sqr.dev, glob.lens, col)
}
\arguments{
\item{glob.sqr.dev}{is the global squared deviance per column.}

\item{glob.lens}{is the global lengths of each column.}

\item{col}{Should be supplied by researcher as a vector of strings
representing the column names they think is present in the data}
}
\value{
named vector of global variance per column.
}
\description{
Function to compute the global variance from global squared deviance and
global lengths
}
