% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_formula.R
\name{init_formula}
\alias{init_formula}
\title{Aggregator function}
\usage{
init_formula(f)
}
\arguments{
\item{f}{an object of class formula or one that can be converted.}
}
\value{
list with formula.
}
\description{
Returns formula in a list
}
