% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_deviance.R
\name{master_deviance}
\alias{master_deviance}
\title{RPC call for the second data loop of the federated GLM}
\usage{
master_deviance(nodes = NULL, master)
}
\arguments{
\item{nodes}{list of results from the individual nodes}

\item{master}{a list of parameters used to compute the GLM}
}
\value{
global updated parameters
}
\description{
RPC call for the second data loop of the federated GLM
}
