% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cox_zph.R
\name{plot_cox_zph}
\alias{plot_cox_zph}
\title{This plots the Cox Zph and the Schoenfeld Residuals...}
\usage{
plot_cox_zph(
  cox_zph,
  resid = TRUE,
  se = TRUE,
  df = 4,
  num_pts = 40,
  xlab = "Time",
  ylab = "",
  lty = 1:2,
  col = 1,
  lwd = 1
)
}
\description{
This plots the Cox Zph and the Schoenfeld Residuals...
}
