% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.data.R
\name{preprocess.data}
\alias{preprocess.data}
\title{Split the dataframe column wise into covariate, censor and time columns.}
\usage{
preprocess.data(df, expl_vars, censor_col, time_col)
}
\description{
Split the dataframe column wise into covariate, censor and time columns.
}
