% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPC_compute_summed_z.R
\name{RPC_compute_summed_z}
\alias{RPC_compute_summed_z}
\title{Compute the aggregate statistic of step 2
sum over all distinct times i
  sum the covariates of cases in the set of cases with events at time i.}
\usage{
RPC_compute_summed_z(
  df,
  subset_rules,
  expl_vars,
  time_col,
  censor_col,
  types = NULL
)
}
\description{
Params:
  df: dataframe
  expl_vars: list of explanatory variables (covariates) to use
  time_col: name of the column that contains the event/censor times
  censor_col: name of the colunm that explains whether an event occured or
              the patient was censored
}
\details{
Return:
  numeric vector with sums and named index with covariates.
}
