% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPC_CT.R
\name{RPC_CT}
\alias{RPC_CT}
\title{Data station/ node side function.}
\usage{
RPC_CT(data, master)
}
\arguments{
\item{data}{Data provided by client.}

\item{master}{This will contain output from `init_formula`, a formula.}
}
\value{
Local instance of Cross tabulation. This computes the instances per
var_cat which belong to the data. In other words, frequency distribution
per categorical variable.
}
\description{
Calculates the local instance of the Cross tabulation.
}
