% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPC_extract_event_times.R
\name{RPC_extract_event_times}
\alias{RPC_extract_event_times}
\title{RPC call for secondary step of Federated Coxzph, extract unique event times.}
\usage{
RPC_extract_event_times(data, time, event)
}
\arguments{
\item{data}{dataframe containing the data, this is automatically supplied
by the node}

\item{time}{this is a string detailing which column of the dataframe the
algorithm should extract the time from}

\item{event}{this is a string detailing which column of the dataframe the
algorithm should extract the event from}
}
\value{
uniuqe event times
}
\description{
RPC call for secondary step of Federated Coxzph, extract unique event times.
}
