% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPC_get_statistics.R
\name{RPC_get_statistics}
\alias{RPC_get_statistics}
\title{Get Node Statistics}
\usage{
RPC_get_statistics(data, col, threshold = 5L, types = NULL)
}
\arguments{
\item{data}{Dataset}

\item{col}{Should be supplied by researcher as a vector of strings
representing the column names they think is present in the data}

\item{threshold}{for disclosure check. Default is 5.}
}
\value{
a list of na, lengths, sums, range and useable rows
}
\description{
This function calculates all the preliminairy statistics used in the summary
algorithm. It calculates number of NA or missing data, length of each column
in the data, sum per column of data given that the data is not a factor,
and the range per column of the data:- if the column is a factor
then it returns a table of values with a disclosure check, otherwise
it simply returns the min and max values. Finally this also returns the
number of useable rows in the data.
}
