% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPC_get_unique_event_times_and_counts.R
\name{RPC_get_unique_event_times_and_counts}
\alias{RPC_get_unique_event_times_and_counts}
\title{Return a dataframe of unique event times}
\usage{
RPC_get_unique_event_times_and_counts(
  df,
  subset_rules,
  time_col,
  censor_col,
  types = NULL
)
}
\description{
Params:
  df: dataframe
  time_col: name of the column that contains the event/censor times
}
\details{
Return:
  dataframe with columns time and Freq
}
