% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPC_get_vars.R
\name{RPC_get_vars}
\alias{RPC_get_vars}
\title{Data station/node side function.}
\usage{
RPC_get_vars(data, master)
}
\arguments{
\item{data}{Data provided by client.}

\item{master}{output from `init_formula`, a formula.}
}
\value{
returns a list with unique categories/factors of the variable in
the data.
}
\description{
This function extracts from the data based on the formula provided the
unique levels/categories/factors.
}
