% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPC_localdev.R
\name{RPC_localdev}
\alias{RPC_localdev}
\title{RPC call for single iteration of nAGQ procedure to find minimized Deviance.}
\usage{
RPC_localdev(data, formula, params, family, nAGQ)
}
\arguments{
\item{data}{Dataframe containing the data which is automatically supplied by
the node.}

\item{formula}{A two-sided linear formula object describing both the fixed
effects and random effect parts of the model. The response parameter must be
on the left hand side of a '~' operator and the independent variables that
follow must be separated with a '+' operator. The random effect term must
be distinguished by a vertical bar '|'.}

\item{family}{Family type, if non is supplied then Gaussian Family is used.}

\item{nAGQ}{Integer Scalar of the number of points per axis for evaluating
the Adaptive Gauss Hermite Quadrature approximation to the log-likelihood.
The default number we take is 10, but this can go up to 25 or as few as 1.
The latter will simply be the Laplace Approximation of the Deviance.}

\item{start}{List of starting values for the Random Effect Term and the
Fixed Effect Term(s). Note that one should invoke the list in the following
manner: start = list(theta = x, fixef = y) where `x` is a single value, and
`y` can be a vector of starting values depending on how many fixed effects
are included in the model.}

\item{len.theta}{Integer Scalar- since the current implementation of the
Adaptive Gauss Hermite Quadrature Scheme used in glmer only allows for one
Random Effect Intercept Term, this has value has to be set to 1. If not then
no optimization will proceed.}
}
\value{
A Single Iteration of the nAGQ approximation of the minimized
Deviance. Along with this, the gradient and hessian are given as attributes.
}
\description{
Executes a single iteration of the glmer Function from the lme4 package to
obtain a deviance, gradient and hessian. For further details on the glmer
function please see `help(lme4::glmer)`.
}
