% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPC_perform_iteration.R
\name{RPC_perform_iteration}
\alias{RPC_perform_iteration}
\title{Compute the three aggretated statistics needed for an iteration}
\usage{
RPC_perform_iteration(
  df,
  subset_rules,
  expl_vars,
  time_col,
  censor_col,
  beta,
  unique_event_times,
  types = NULL
)
}
\description{
Params:
  df: dataframe
  expl_vars: list of explanatory variables (covariates) to use
  time_col: name of the column that contains the event/censor times
  censor_col: name of the colunm that explains whether an event occured or
              the patient was censored
  beta: vector of beta coefficients (length(beta) == length(expl_vars))
  times: vector of *globally* unique event times
}
\details{
Return:
  list containing aggretated statistics
}
