% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPC_ratio.R
\name{RPC_ratio}
\alias{RPC_ratio}
\title{node-side function used further in calculatin of Hazard Ratio.}
\usage{
RPC_ratio(data, time, event, coxfit)
}
\description{
node-side function used further in calculatin of Hazard Ratio.
}
