% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPC_sqr_dev.R
\name{RPC_sqr_dev}
\alias{RPC_sqr_dev}
\title{Calculate Squared Deviance.}
\usage{
RPC_sqr_dev(data, col, glob.mean)
}
\arguments{
\item{data}{Dataset}

\item{col}{Should be supplied by researcher as a vector of strings
representing the column names they think is present in the data}

\item{glob.mean}{This is calculated earlier in the algorithm, global mean
of the combined datastation(s) (nodes).}
}
\value{
Vector of squared deviance per column in the Data or NaN if the
data is populated entirely by NA
}
\description{
This function is the precursor to getting the summed squared deviance
across all the datastation(s) (node). This is then used to calculate the
variance and following that, the standard deviation.
}
