% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_vars_list.R
\name{sort_vars_list}
\alias{sort_vars_list}
\title{server-side function but should probably be nodeside...}
\usage{
sort_vars_list(nodes = NULL, coxfit = NULL)
}
\description{
server-side function but should probably be nodeside...
}
