% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_data.R
\name{subset_data}
\alias{subset_data}
\title{version: 1.0
Data filter function.}
\usage{
subset_data(data, subset_rules)
}
\arguments{
\item{data}{Data provided by client.}

\item{subset_rules}{A list of subset rules given by the user. When providing 
multiple subsetting steps the subsets will be executed in the order as given 
in the list. 

Example of a subset rules json file:
[
 {
   "subset": "site==6"
 },
 {
   "subset": "site %in% c(6,7)"
 },
 {
   "subset": "age>30 & age<=70"
 }
]}
}
\value{
Either, the subsetted data to be used in the rest of the analysis, 
or, for privacy preserving reasons, an empty dataframe (with the column 
names) if the subsetted data has less than N rows.
}
\description{
This function will be called at the start of the analysis, and executed on 
the data of every node. Has to be implemented in every RPC call.
}
