% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_categories.R
\name{variable_categories}
\alias{variable_categories}
\title{Aggregator side function.}
\usage{
variable_categories(nodes, master)
}
\arguments{
\item{nodes}{This contains the categories from each node. It is a list.}

\item{master}{List containing output from `init_formula`. Formula.}
}
\value{
Returns the master object with appended item, var_cat which is
the global unique categories.
}
\description{
Collect all the local instances of unique variables and sort them to a
global list of unique categories.
}
