% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeln.R
\name{writeln}
\alias{writeln}
\title{Write a string to STDOUT without the standard '[1]' prefix.}
\usage{
writeln(..., sep = " ")
}
\arguments{
\item{...}{Strings to write to STDOUT}

\item{sep}{Separator to use between strings. Defaults to a single space}
}
\description{
Write a string to STDOUT without the standard '[1]' prefix.
}
\examples{
writeln("Hello, world!")
}
