% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dglm.R
\name{dglm}
\alias{dglm}
\title{Federated GLM}
\usage{
dglm(
  client,
  formula,
  dstar = NULL,
  types = NULL,
  family = gaussian,
  tol = 1e-08,
  maxit = 25,
  organizations_to_include = NULL,
  subset_rules = NULL
)
}
\arguments{
\item{client}{vtg::Client instance, profided by the node}

\item{formula}{an object of class formula (or one that can be coerced
to that class: a symbolic description of the model to be fitted.
E.g.: dependant_variable ~ explanatory_variable(i) + ...}

\item{dstar}{name of the dstar sensor column (e.g. expected deaths).
Only applicable when using the glm relative survival model with
Poisson regression)}

\item{types}{types of the columns that are used in the formula}

\item{family}{family type, Gaussian is used by default}

\item{tol}{tolerance level}

\item{maxit}{maximum number of iterations}
}
\value{
A GLM model in a dict format. To convert it to a GLM model which
 can be used by R, use the `vtg.glm::as.GLM(output)` to convert it.
}
\description{
Federated GLM
}
\author{
Cellamare, M.

Martin, F.

van Gestel, A.
}
