% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_family.R
\name{get_family}
\alias{get_family}
\title{translates family string input to R family object}
\usage{
get_family(family, dstar = NULL)
}
\arguments{
\item{family}{sting identifyer of the family to be returned}
}
\value{
R family object
}
\description{
translates family string input to R family object
}
