% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_beta.R
\name{master_beta}
\alias{master_beta}
\title{Combine the results from the first RPC call and update the betas
accordingly}
\usage{
master_beta(partials1, family, dstar, master = NULL)
}
\arguments{
\item{master}{a list of parameters used to compute the GLM}

\item{nodes}{list of results from the individual nodes}
}
\value{
global updated parameters
}
\description{
Computes: X^T W X and X^T W z from the partials. Then computes the
updated betas by multiplying these.
}
