% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_beta.R
\name{master_beta}
\alias{master_beta}
\title{Combine the results from the first RPC call and update the betas
accordingly}
\usage{
master_beta(partials, family, dstar)
}
\arguments{
\item{partials}{result of rpc_node_beta, partial results from the
individual nodes}

\item{family}{family type}

\item{dstar}{name of the dstar sensor column (e.g. expected deaths).
Only applicable when using the glm relative survival model with
Poisson regression)}
}
\value{
global updated parameters
}
\description{
Computes: X^T W X and X^T W z from the partials. Then computes the
updated betas by multiplying these.
}
