% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPC_node_beta.R
\name{RPC_node_beta}
\alias{RPC_node_beta}
\title{RPC call for the first data loop of the federated GLM}
\usage{
RPC_node_beta(
  data,
  subset_rules,
  formula,
  family,
  first_iteration,
  coeff,
  dstar = NULL,
  types = NULL,
  weights = NULL
)
}
\arguments{
\item{data}{dataframe containing the data, this is automatically
supplied by the node}

\item{formula}{an object of class formula (or one that can be coerced
to that class: a symbolic description of the model to be fitted.
E.g.: dependant_variable ~ explanatory_variable(i) + ...}

\item{family}{family type, Gaussian is used by default}

\item{first_iteration}{boolean to indicate if this is the first
iteration}

\item{coeff}{beta coefficients from previous iteration}

\item{dstar}{name of the dstar sensor column (e.g. expected deaths).
Only applicable when using the glm relative survival model with
Poisson regression)}

\item{types}{types of the columns that are used in the formula}

\item{weights}{an optional vector of ‘prior weights’ to be used in the
fitting process. Should be NULL or a numeric vector}
}
\value{
GLM partials
}
\description{
RPC call for the first data loop of the federated GLM
}
